parser v0.2.1-main.sha.1171107+gbad.14: reimplement HTMLParser
The previous implementation would mishandle other tags
like <b> and <i>, which was why I added them
in previous patches, and it also lost text before
single <br> tags.

Finally, single line breaks are now converted
to spaces and not dropped as before.

Not sure if there are any downsides of this
implementation, but we will have to test and see!