Start rewriting map_schema with in-drawio URI masks
Also prepping to bundle everything in one package,
so transitioned to a standalone Python script here.

This is a halfway version, just committing so that
it is saved here - this work is unfinished.

The previous map_schema Jupyter notebook should still be compatible
with the previous version of general graphs and tables, though, but they are deprecated now.