Update map_schema.py to support supplemental graphs
Important! I observed that adding duplicate triples
to the same rdflib graph is possible, and this leads
to duplicated maps in RML, which in turn leads
to incorrect generations.

Thus, it is important that triples are not duplicated
in the main and supplemental graphs - and this is why
I have cleaned the main graph in the parent commit.