Remap authority using updated graph and mapper
Important! As we can see from the mapped triples,
AgentControlRelation entities are generated for *each*
Authority record, even if there is no actual controlling
agency. In this case, 13-14 not necessary nonexistent
relation entities are generated, which will make the output
graph heavier and messier.

In fact, this is the expected behavior. Given that these
are generated by the mapper, I do not know at this point
if there are any mechanisms in RML that allow one to selectively
limit generations.

Thus, I think a solution for these will be removing them
at postprocessing using a SPARQL query, e.g., see
attached "Remove_phantom_AgentControlRelation.rq"