Fix some bugs in map schema script from recent commits
URIs would be generated as literals for rr:constant
even where URIRef was valid; urllib parse escaping
was unsafe which invalidated the protocol part;
and subject mnemonic triples were generated
for de novo entities; and a couple minor tweak.

The RML looks good to me at this point, and conversion
works as well.