Fix bug in map_schema that wouldn't create subject maps for increments
Whoa, this was a wicked one. The RML looks like it should be working
now, but the head 101 dataset apparently does not contain
any agent control relations, so cannot check that,
but PREV and SUC field appear to have been added to the
mapped file, and no triples seem to have been deleted.

I will run the conversion of the entire authority dataset in next commit.