Sort-of-implement removal of phantom entities in code
Also commented out all postprocessing
because it is not needed at the moment -
agent control relations should get resolved automatically.

This is "sort of" because even though the code seems to be working
correctly, it does not actually remove all of the phantoms here:
Although it looks like it does remove empty agent control relations
without any postprocessing this time, cases like the original test case
REFA is not worked through because that subject map really depends
on HEADING, not REFA (bc REFA is used as rico:normalizedValue),
so this is not incorportated in the RML.

Although it is definitely possible to scan for those cases within rml_schema
and add them into a sort of "dependent_mnemonic" column
of the dataframe, I am not sure if this will be necessary with the new
all-dataset KB naming scheme discussed today, because that will probably
depend on the corresponding mnemonic only.

Thus, I will revisit the graph, changing the URI format, and will try again.
The FnO code itself seems to work correctly.