Explore removing duplicate incremented objects
It looks like now this works the way it is expected
to work: increments in one node do not affect
increments in the other node, which is achieved
through separate increment disaggregation in objects.

Thus, the most appropriate approach at this point
seems to remove rico:textualValue from VAR and PAR
because they are not essential, but I will keep this
intricate aspect in mind and will modify the increment
behavior later on if needed.

So I will rework the general auth graph again in next commit.