Fix the issue of non-unique AgentControlRelation URIs
Although the URIs seem to be unique now, the mapped triples
are still buggy - see the bottom of this commit body.

This also reworks UUID logic, making it somewhat cleaner
and better documented.

The reason why AgentControlRelation URIs are now unique
is because while the right mnemonics should ensure
the uniqueness of the row, the TriplesMap name,
based on which the UUID is generated, ensures column uniqueness.

I have checked this with AUTHORITY.csv, and it seems to work fine now
on the surface. However, upon closer checking, it looks like
object triples are incorrectly assigned to all of the incremented
fields, e.g., see this from the sample:

<https://data.archives.gov.on.ca/KB/AgentControlRelation/Ontario.%20Dept.%20of%20Colleges%20and%20Universities/Ontario%20Committee%20on%20Student%20Awards/19df1c2a-a5f7-5f7e-9f11-aebb1b4a371a>
  a rico:AgentControlRelation;
  rdfs:label "Ontario. Dept. of Colleges and Universities (AgentControlRelation)";
  rico:hasBeginningDate <https://data.archives.gov.on.ca/KB/Date/1966>, <https://data.archives.gov.on.ca/KB/Date/1971>,
    <https://data.archives.gov.on.ca/KB/Date/1972>;
  rico:hasEndDate <https://data.archives.gov.on.ca/KB/Date/1971>, <https://data.archives.gov.on.ca/KB/Date/1972>,
    <https://data.archives.gov.on.ca/KB/Date/1974>;
  rico:relationHasTarget <https://data.archives.gov.on.ca/KB/CorporateBody/Ontario%20Committee%20on%20Student%20Awards> .

In the data, this relation only took place from 1972 to 1974
whereas all the other dates come from the other relations.
I will be fixing this now.