Start revising general Description graph according to latest URI scheme
Also for the first time separated Description general graph
from Listings general graph (I will work on the latter later).

It is of note that CreationRelation has not been updated and not sure yet
if I should continue keeping it as a separate supplemental graph
or, rather, move it back to the main graph given that there is more space now.