Combine all ADD general graphs back into one graph
I reflected on this and realized that custom logic will still be necessary, e.g., {REF} to process
the URI of the main node depending on the value of REF_FILE for Listings,
so it does not make much sense to separate.

Also, it is quite easy to switch the input file using the command line logic already implemented,
so a different graph is also not necessary.

Finally, the good news is that the drawio parser seems
to parse unconnected graphs correctly, which is great
because I can now model different parts of the graph
as separate graphs but all in one graph file, e.g., like
it is now with the "main" graph and creation relations.