Add 'fake' columns to listings sample header and convert some records
Just to see if this works on that file, too. It seems to work!

Though having the converted file replace the description mapping
is probably not the expected behavior, so I could modify the script
to also add the source title to the mapped directory name, but let me see.