parser v0.2.1-main.sha.1171107+gbad.17: allow custom types and subproperties
Might be buggy but seems to work. Overall
the approach is not at all stable (e.g., setting
Domain or Range for properties is unsupported
yet, and there may be bugs with custom classes),
but I need this to make this work at this point.

This also reworks general Authority graph
to account for changes, as well as adds
a few missing fields to that graph.

In next steps, I will add the functionality
for custom properties to the map_schema script.