Create riconvert and post commit hook to keep track of versions
The hook is not totally safe for modifying my file and especially
removing the first two lines, but let it be for now.

The idea is to keep this info in main branch file at all times
in case somebody used them, so I can always backtrack.
I do not feel like I can totally rely on the releases because
someone could still want to use the original file directly from repo,
so let us try this for now.

The riconverter file will be a bundled script with everything
from drawio converter to rml mapper in.