Implement REFD_FILE
All seems to work!

Although it should be noted that CSV header must be
the same for Description and Listings source files
for this to work, so I added missing columns as illustrated
in updated description tailshuf 100.

The missing fields were noted by rmlmapper because
it gave an error, so I just added all columns it showed.