Remove REF_ADD for Listings with REF_FILE
I have done that, although I note that some
listings only contain a number in that field,
e.g., just 1 instead of full reference code.

However, at this point, it really seems the best
approach because REF_ADD is referenced
via directlyIncludedIn anyways, so won't be
lost even if it's not in REF_FILE.
But it would be problematic if URIs
contain duplicative codes.