Enable fully controlled generation of rdfs:label
First, we add --label-disable in bash scripts to prevent the creation of duplicates.

Second, we add an update to map schema
that will autogenerate rdfs:label
if missing in its own way but preserve any rdfs:label that comes
from drawio file.

This also disables the awkward hardcoded rdfs:label generation
from TITLE, so this will later on specified in the drawio itself.