Reimplement REFD_FILE so that it is now only REFD or REF_FILE
As decided at a team meeting today, REF_FILE should ideally contain
a full ref code, including its REF_ADD part, so we decided that
this is what we implement, and then we isolate the records that don't follow this
and treat them separately. Many records still do follow the right format.