Fix non-disaggregated GMD and simplify general ADD graph
So this was an obvious error for GMD.

For OFFICEABC, however, which was also not generated correctly, the situation is wicked-er
because I observed that an incremented OFFICEABC_2 containing URI
has failed to generate for rico:thingIsSourceOfRelation.

And when I think of it, it is totally understandable because how disaggregation by increment
works now for objects is that it simply carries forward the increment number
from subject, and only if this number is not set, is disaggregation logic
invoked for object, but otherwise subject's increment is simply reused.

Also, it is of note that mnemonic (extracted from object) is used for increments,
whereas in "KB/CreationRelation/{REFD_FILE}/{OFFICEABC_1..2}/{UUID_OFFICEABC},
REFD_FILE is the mnemonic (leftmost term), not OFFICEABC_1..2!
As this is documented even in map_schema.log warning.
Therefore, it is quite clear why this is not disaggregated.

Anyhow, I just decided to make things simpler and made OFFICEABC 1 and 2
into two separate graphs - this ensures that all works as expected
whereas for simpler cases, like GMD_1..8, automatic disaggregation
can still be safely used.