Complete full description conversion
Some bugs related to how preprocessing was done
are obvious, e.g., "Mc" in "McDonald" or alike is separated out,
but we can review this and think what we can do about it.
LLM, even a locally run small model, would probably be best.

As can be seen from the log, DtypeWarning is raised
but did not seem to affect execution time significantly,
so I think I can let this pass for now.