Implement pull of Authority DATEEX from a new export
Not committing the export because no need to stack up
a lot of files in this repo.

Also needed to rerun parse-sh to update the Authority graph
because rdfs:label with "rr:constant ..." from older edition
would not parse correctly now (not exactly sure why).

I have checked preprocessed authority head 6 and tailshuf 100 against
the original ones, and they seem to do fine and only replace
the values that should be replaced.

Will still need to re-check on the full Authority dataset.