Add a basic rdflib sparql endpoint for testing
This works overall but has issues with
content negotiation - specifically,
service_description attribute is not supported
herem and thus I could not change the default
content type this sends, and thus could
not link this to Trifid that somehow expects
a different flavor of xml. I will probably have
to fix this behavior within Trifid.

But the endpoint itself works! But is quite slow on my PC when I tried
running it with the actual full Description dataset (postprocessed
and serialized as N-triples):

The start time took maybe a couple of minutes (did not time it),
and here is what I got for SELECT (COUNT(*) AS ?count)

1 result in 19.127 seconds

(which did return the right count of 785711)