Push prerelease version of gbad vsce v0.0.2
This is still prerelease because does not fully do what it's intended to:

- No useful placeholders (e.g., UUIDs, current date)
- All prefixes are lost; so I tried fixing this but somehow
the extension does compile but validate and serialize button
does not do anything anymore, so I had to comment out
and will debug.
- Output format selection in settings does not work - always TriG is produced.

Anyway, I still decided to push this because the validation itself
seems to work: e.g., I tested manually a few valid and invalid cases,
and it serialized successfully valid turtle, trig, nt, and when there is
a validation error, it alerts the place where in the doc it is.