[MAJOR] Redefine what UUIDs are generated from
Now a more straightforward approach is used
whereby each rr:TriplesMap URI in Mapping namespace
is used as context.

Important! UUID for any literals are now generated
from the linked subject, which makes sense
because literals do not get a URI nor triplesmap.

This also fixes logic that prevented UUID in literals
from being generated.

This also fixes prefix and ontology URI in parse auth and add -
this apparently was a long standing but and only
discovered now.

Upon review of RML, it seems to correctly contain
UUIDs for all agent control relations and their labels -
same UUID for each pair (28 in total).