Add custom rdfs:label for RAD entity in general auth graph
This seems to work - even just when writing without
RML syntax, this get converted to an RML literal, which is good.

However, I have noticed that map_schema did not process
the complex domains from schema auth and add, which is
understandable because an empty node is used - just needs a fix.