Move add suppl graphs to postprocessing
This should fixes the issue with passing schema
suppl graphs through rml - see attached.

However, I had to include this into postprocessing
to keep the logic, and it makes sense as it is
in fact postprocessing after RML mapping.

Also, added triples are not included in the count
anymore in logs, which kind of makes sense too
as these do not come from the conversion.