Pin dependencies for GBAD project
A Conda and Poetry configs are now available,
which can be used together (recommended)
or in separation.

The versions of openjdk and python are pinned
and are inspired by a real-life reproduction attempt
at the Archives of Ontario in December 13, 2024
(see subdir "riconverted_general_authority_to_ric-o_model_2024-11-25_pz")
for some more insights into that experiment.

Latest versions of rdflib and pandas are used here,
which is important to note.

Makefile can be used to simplify installing and uninstalling.

The main requirements.txt file remains unchanged.