Fix relative URI resolution issue
Thanks Claude Sonnet 4! This was a wicked one.

For some reason relative URIs were
resolved upon serialization by post
processing script relative to the current
file.

I even tried to set '-b' arg for rmlmapper-7.0.0-r374-all
(by the way, which is not supported by rmlmapper-4.0.0, which
is an earlier version I had to use to test on 2024-12-13 at the Archives
of Ontario because the closest available version of Java was openjdk=8.0.282),
but it still woudn't produce the @base statement in mapped.ttl,
which led to the incorrect parsing (when I set @base manually,
everything would parse correctly).

But luckily now with publicID set, this works.

I had to set up a 'next' environment to do all that therefore
to be sure to add a newer version of Java (openjdk=21.0.6 worked).
I also updated both the linux and macos test_rg_1-429 scripts
(they both work but on linux the conda bash hook returns some error
due to trying to run openjdk_activate and deactivate - this doesn't
affect the script itself).

I also tried to map using pyrml for the first time - dep is here but will
commit the code to experiment next (it failed to execute).