[MAJOR] Add support for relative URI in rr:constant
That is, it is OK now to start with slash or hash in drawio.

This gives the flexibility not to hardcode the base URI in the
drawio graphs, which is quite valuable.

See attached updated General ADD drawio with one node
changed as an illustration.

I checked this with tests/scripts/linux/test_rg_1-429.sh
and seems to work well. I will attach results later because
I also implemented some other changes.