Add support for leading slash/hash in drawio template
See attached drawio with a few nodes updated as an
example, even with #KB/Title but not for real but
as an example that hash also works.

So for literal nodes (ovals), leading slash or hash
is preserved, and for URI nodes, it is attached
to base URI (as can be seen in the map_rml update).

When trailing slash or has was preserved in base URI,
this lead to duplication.

This seems to work well now when tested with test_rg_1-429.sh