[wip] explore what we can do for local ontologies
So with some edits (attached), this seems to
work more or less satisfactorily, but the thing
is that for local ontology files, the same path is used
as prefix_iri, which doesn't allow for matching. Should
I add a separate option for ontology path specifically?
Forcing users to have their ontology online doesn't
sound like a good idea.